examples/ocf_client: Remove the parsing of responses
For the purposes of the ocf_client, if the user is able to read the
JSON-like representation of the server responses, we don't need to
parse the responses.

This depends that the CBOR_Dump() function prints the content of CBOR
payloads, which is enabled by commit
"cbor: Enable CBOR payload debugging (DO NOT APPLY)".

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>